refactor(backend): split agentcore-memory into codec/mapper, dedupe record mapping
Extract the wire-format decoding (parseBlobPayload, convertToMessageContents and
their block/message types) into services/memory/content-codec.ts, and the
long-term-record mapping into services/memory/record-mapper.ts. The service
(774 -> 458 lines) now stays in domain terms and re-exports the moved symbols so
existing importers are unaffected.

- listMemoryRecords / retrieveMemoryRecords shared a ~25-line summary->record
  block each; both now call the single pure mapMemoryRecord().
- ResourceNotFoundException->empty handling collapses into one
  withEmptyOnNotFound() policy helper.
- Remove the index<2 'fullRecord' debug logs that dumped entire memory records
  (which can carry user content) to the production logs.

mapMemoryRecord is covered by a focused unit test; full backend suite green.
